pyproject.toml
[tool.poetry]
name = "lx_music_api_server"
version = "2.0.0"
description = "一个适配 LX Music 的 API 后端实现"
authors = ["helloplhm-qwq", "lerdb", "Folltoshe"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.10.10"
pycryptodome = "^3.22.0"
ujson = "^5.10.0"
requests = "^2.32.4"
colorlog = "^6.9.0"
Pygments = "^2.19.1"
xmltodict = "^0.14.2"
pillow = "^10.4.0"
mutagen = "^1.47.0"
colorama = "^0.4.6"
ruamel-yaml = "^0.18.14"
pybind11 = "^2.13.6"
redis = "^6.1.1"
hiredis = "^3.1.0"
pystray = "^0.19.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"